refactor: add network
