+++ tests/selectors/network.selectors.ts
@@ -1,9 +1,10 @@
 export enum NetworkSelectors {
+  NetworkPageReady = 'network-page-ready',
   NetworkName = 'network-name',
   NetworkStacksAddress = 'network-stacks-address',
   NetworkBitcoinAddress = 'network-bitcoin-address',
   NetworkKey = 'network-key',
-  BtnAddNetwork = 'btn-add-network',
+  AddNetworkBtn = 'add-network-btn',
   ErrorText = 'error-text',
   EmptyNameError = 'Enter a name',
   EmptyStacksAddressError = 'Enter a valid Stacks API URL',
