+++ tests/selectors/settings.selectors.ts
@@ -17,10 +17,10 @@ export enum SettingsSelectors {
   LockListItem = 'settings-lock',
   EnterPasswordInput = 'password-input',
   UnlockWalletBtn = 'unlock-wallet-btn',
-  BtnAddNetwork = 'btn-add-network',
+  AddNewNetworkBtn = 'add-new-network-btn',
   ShowSecretKeyBtn = 'show-secret-key-btn',
   GetSupportMenuItem = 'get-support-menu-item',
-  SettingsMenuBtn = 'settings-menu--trigger',
+  SettingsMenuBtn = 'settings-menu-trigger',
   SwitchAccountTrigger = 'switch-account-trigger',
   SwitchAccountMenuItem = 'switch-account-menu-item',
   SwitchAccountItemIndex = 'switch-account-item-[index]',
